refact: Refactors contract hooks for modularity and maintainability:
 - Reorganizes contract-related logic into separate context and provider files to improve code structure and readability.
 - Adds enhanced error handling, type definitions, and new functionality for fetching questions and answers from the contract.
 - Updates TypeScript configuration and Vite aliases to support new types directory.
