--- frontend/react/src/hooks/useContract/contractContext.ts
@@ -1,20 +1,0 @@
-import { createContext, useContext } from 'react'
-import { useContract as useStarknetContract } from "@starknet-react/core"
-
-interface ContractContextType {
-  contract: ReturnType<typeof useStarknetContract>['contract']
-  contractReady: boolean
-  isConnected: boolean | undefined
-  address: string | undefined
-  abiError: string | undefined
-}
-
-export const ContractContext = createContext<ContractContextType | undefined>(undefined)
-
-export function useContractContext() {
-  const context = useContext(ContractContext)
-  if (context === undefined) {
-    throw new Error('useContractContext must be used within a ContractProvider')
-  }
-  return context
-}
