+++ frontend/react/src/hooks/useContract/errors.ts
@@ -1,0 +1,7 @@
+// Error messages
+export const ERROR_MESSAGES = {
+  CONTRACT_NOT_INITIALIZED: "Contract not initialized / Wallet not connected",
+  QUESTION_NOT_FOUND: "Question not found",
+  FETCH_QUESTION_FAILED: "Failed to fetch question",
+  FETCH_ANSWERS_FAILED: "Failed to fetch answers"
+}
