+++ frontend/react/src/hooks/useContract/index.ts
@@ -1,123 +1,12 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
-import { useState, useCallback } from "react"
-import { useAccount } from "@starknet-react/core"
-import { ContractAnswer, ContractQuestion } from "../../services/contract"
-import { contractQuestionToFrontend, contractAnswerToFrontend } from "../../utils/contractTypeMapping"
-import { Question, Answer } from "../../pages/AnswerPage/types"
-import { formatters } from "@utils/formatters"
-import { useContractContext } from "./contractContext"
-
-
-interface ContractState {
-  isLoading: boolean
-  error: string | null
-  transactionHash: string | null
-}
+import { useContext } from "react"
+import { ContractContext } from "./contract.context"
 
 export function useContract() {
-  const { isConnected } = useAccount()
-  const { contract, contractReady, abiError } = useContractContext()
+  const context = useContext(ContractContext)  
 
-  const [questionState, setQuestionState] = useState<ContractState>({
-    isLoading: false,
-    error: null,
-    transactionHash: null
-  })
-
-  const [answersState, setAnswersState] = useState<ContractState>({
-    isLoading: false,
-    error: null,
-    transactionHash: null
-  })
-
-
-
-  // Fetch question from contract
-  const fetchQuestion = useCallback(async (questionId: string): Promise<Question | null> => {
-    if (!contract || abiError) {
-      setQuestionState({ isLoading: false, error: abiError || "Contract not initialized / Wallet not connected", transactionHash: null })
-      return null
-    }
-
-    setQuestionState({ isLoading: true, error: null, transactionHash: null })
-
-    try {
-      const result = await contract.get_question(formatters.numberToBigInt(Number(questionId))) as ContractQuestion
-
-
-
-      if (!result.description || !formatters.bigIntToAddress(result.id)) {
-        setQuestionState({ isLoading: false, error: "Question not found", transactionHash: null })
-        return null
-      }
-
-      const question = contractQuestionToFrontend(result)
-
-      setQuestionState({ isLoading: false, error: null, transactionHash: null })
-      return question
-    } catch (error) {
-      const errorMessage = error instanceof Error ? error.message : "Failed to fetch question"
-      setQuestionState({ isLoading: false, error: errorMessage, transactionHash: null })
-      return null
-    }
-  }, [contract, abiError])
-
-  // Fetch answers for a question
-  const fetchAnswers = useCallback(async (questionId: string): Promise<Answer[]> => {
-    if (!contract || abiError) {
-      setAnswersState({ isLoading: false, error: abiError || "Contract not initialized / Wallet not connected", transactionHash: null })
-      return []
-    }
-
-    setAnswersState({ isLoading: true, error: null, transactionHash: null })
-
-    try {
-      const [contractAnswers, correctAnswerId] = await Promise.all([
-        contract.get_answers(formatters.numberToBigInt(Number(questionId))) as Promise<ContractAnswer[]>,
-        contract.get_correct_answer(formatters.numberToBigInt(Number(questionId))).catch(() => BigInt(0)) as Promise<bigint>
-      ])
-
-      const answers = contractAnswers.map((contractAnswer: ContractAnswer) =>
-        contractAnswerToFrontend(
-          contractAnswer,
-          contractAnswer.id === correctAnswerId
-        )
-      )
-
-      setAnswersState({ isLoading: false, error: null, transactionHash: null })
-      return answers
-    } catch (error) {
-      const errorMessage = error instanceof Error ? error.message : "Failed to fetch answers"
-      setAnswersState({ isLoading: false, error: errorMessage, transactionHash: null })
-      return []
-    }
-  }, [contract, abiError])
-
-
-
-  // Clear errors
-  const clearQuestionError = useCallback(() => setQuestionState(prev => ({ ...prev, error: null })), [])
-  const clearAnswersError = useCallback(() => setAnswersState(prev => ({ ...prev, error: null })), [])
-
-  return {
-    // Question fetching
-    fetchQuestion,
-    questionLoading: questionState.isLoading,
-    questionError: questionState.error,
-    clearQuestionError,
-
-    // Answers fetching
-    fetchAnswers,
-    answersLoading: answersState.isLoading,
-    answersError: answersState.error,
-    clearAnswersError,
-
-
-    // General
-    isConnected,
-    contractReady,
-
-    contract,
-
+  if (!context) {
+    throw new Error("useContract must be used within a ContractProvider")
   }
+
+  return context
 }
